Create sync endpoint and send offline data
A major issue with the Network.addListener function was that I wrote it outside of an effect, which led to this function being called multiple times instead of doing that only once. The result was that multiple requests were made for synchronizing the offline data with the one on the server.